requirements.txt
Django==1.11
PyYAML==3.12
six==1.10.0
uritemplate==3.0.0
djangorestframework==3.7.3pip